backend/requirements.txt
# Core
Flask==3.1.0
flask-cors==5.0.1

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
psycopg2-binary==2.9.10

# Authentication
Flask-JWT-Extended==4.7.1
bcrypt==4.3.0

# Redis & Task Queue
redis==5.2.1
rq==2.1.0

# Utilities
python-dotenv==1.0.1
marshmallow==3.25.1

# Docker (for judge)
docker==7.1.0

# Dev
gunicorn==23.0.0
